+++ katana/sdk/component.py
@@ -14,6 +14,7 @@ from __future__ import absolute_import
 import logging
 
 from ..errors import KatanaError
+from ..logging import INFO
 from ..logging import value_to_log_string
 from ..schema import SchemaRegistry
 from ..utils import Singleton
@@ -162,7 +163,7 @@ class Component(object):
         self._runner.set_callbacks(self._callbacks)
         self._runner.run()
 
-    def log(self, value):
+    def log(self, value, level=INFO):
         """Write a value to KATANA logs.
 
         Given value is converted to string before being logged.
@@ -171,4 +172,4 @@ class Component(object):
 
         """
 
-        self.__logger.debug(value_to_log_string(value))
+        self.__logger.log(level, value_to_log_string(value))